rename azure file
